pyproject.toml
[project]
name = "sql-murder-mistery-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "python-dotenv>=1.2.2",
]
